Rename package to com.esri.logger.khronicle
