--- src/main/java/com/esri/logger/appender/PrintlnAppender.kt
@@ -1,26 +1,0 @@
-// Copyright 2025 Esri
-// 
-// Licensed under the Apache License Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-// 
-//     http://www.apache.org/licenses/LICENSE-2.0
-// 
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-package com.esri.logger.appender
-
-import com.esri.logger.encoder.Encoder
-import org.slf4j.event.LoggingEvent
-
-class PrintlnAppender : Appender {
-  override var encoder: Encoder? = null
-
-  override fun append(event: LoggingEvent) {
-    encoder?.let { println(it.encode(event).decodeToString()) } ?: { println(event.message) }
-  }
-}
