+++ src/main/java/com/esri/logger/khronicle/FileExtensions.kt
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package com.esri.logger
+package com.esri.logger.khronicle
 
 import androidx.core.text.isDigitsOnly
 import java.io.File
