+++ src/main/resources/META-INF/services/org.slf4j.spi.SLF4JServiceProvider
@@ -1,1 +1,1 @@
-com.esri.logger.ServiceProvider
+com.esri.logger.khronicle.ServiceProvider
