+++ src/test/java/com/esri/logger/khronicle/FileExtensionTests.kt
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package com.esri.logger
+package com.esri.logger.khronicle
 
 import io.mockk.every
 import io.mockk.mockk
