+++ src/test/java/com/esri/logger/khronicle/LoggerBuilderTest.kt
@@ -12,9 +12,9 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package com.esri.logger
+package com.esri.logger.khronicle
 
-import com.esri.logger.appender.TestAppender
+import com.esri.logger.khronicle.appender.TestAppender
 import org.hamcrest.MatcherAssert.assertThat
 import org.junit.Assert.*
 import org.junit.Test
