+++ src/test/java/com/esri/logger/khronicle/PatternEncoderTest.kt
@@ -12,9 +12,9 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package com.esri.logger
+package com.esri.logger.khronicle
 
-import com.esri.logger.encoder.PatternEncoder
+import com.esri.logger.khronicle.encoder.PatternEncoder
 import java.util.TimeZone
 import org.junit.Assert.*
 import org.junit.Before
