+++ src/test/java/com/esri/logger/khronicle/appender/TestAppender.kt
@@ -12,30 +12,16 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package com.esri.logger.appender
+package com.esri.logger.khronicle.appender
 
-import android.util.Log
-import com.esri.logger.encoder.Encoder
-import org.slf4j.MarkerFactory
-import org.slf4j.event.Level
+import com.esri.logger.khronicle.encoder.Encoder
 import org.slf4j.event.LoggingEvent
 
-class LogcatAppender : Appender {
-
-  val defaultMarker = MarkerFactory.getMarker("None")
-
+class TestAppender : Appender {
   override var encoder: Encoder? = null
+  var output: ((LoggingEvent) -> Unit)? = null
 
   override fun append(event: LoggingEvent) {
-    val message = encoder?.encode(event)?.decodeToString() ?: event.message
-    val marker = event.markers?.firstOrNull() ?: defaultMarker
-    when (event.level) {
-      Level.ERROR -> Log.e(marker.name, message)
-      Level.WARN -> Log.w(marker.name, message)
-      Level.INFO -> Log.i(marker.name, message)
-      Level.DEBUG -> Log.d(marker.name, message)
-      Level.TRACE -> Log.v(marker.name, message)
-      else -> Log.v(marker.name, message)
-    }
+    output?.let { it(event) }
   }
 }
