+++ src/test/resources/assets/logger-config.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0"?>
 <configuration xmlns="https://tony19.github.io/logback-android/xml">
-    <appender name="stdout" class="com.esri.logger.appender.TestAppender">
+    <appender name="stdout" class="com.esri.logger.khronicle.appender.TestAppender">
         <encoder>
             <pattern>%message</pattern>
         </encoder>
